+++ packages/svelte/ds-app-launchpad/src/lib/components/SearchBox/common/SearchButton/index.ts
@@ -1,0 +1,2 @@
+export { default as SearchButton } from "./SearchButton.svelte";
+export type { SearchButtonProps } from "./types.js";