+++ packages/svelte/ds-app-launchpad/src/lib/components/SidePanel/types.ts
@@ -1,0 +1,10 @@
+/* @canonical/generator-ds 0.10.0-experimental.2 */
+
+import type { ModalProps } from "../Modal/index.js";
+
+export type SidePanelProps = ModalProps;
+
+export type {
+  ModalMethods as SidePanelMethods,
+  ModalTriggerProps as SidePanelTriggerProps,
+} from "../Modal/index.js";